pyproject.toml
[project]
name = "detecting-attacks"
version = "0.1.0"
description = "AI Safety Protection Layer for LLM Models"
requires-python = ">=3.11"
dependencies = [
    "flask>=2.3.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "accelerate>=0.20.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "huggingface-hub>=0.15.0"
]

[project.scripts]
safety-app = "detecting_attacks.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/detecting_attacks"]
